face-attendance-system/preprocessing/requirements.txt
torch
torchvision
Pillow
imagehash
opencv-python
scikit-learn
matplotlib
numpy
facenet-pytorch
